+++ src/EventIDs.cpp
@@ -311,6 +311,8 @@ const wxWindowIDRef EventIDs::menu_jumptoerror(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_math_as_1D_ASCII(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_math_as_2D_ASCII(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_math_as_graphics(wxWindow::NewControlId());
+const wxWindowIDRef EventIDs::internalRepresentation(wxWindow::NewControlId());
+const wxWindowIDRef EventIDs::wxMathML(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_logexpand_false(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_logexpand_true(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_logexpand_all(wxWindow::NewControlId());